requirements.txt
python-telegram-bot==20.8
discord.py==2.3.2
